+++ include/HD44780_LCD_Print.hpp
@@ -14,14 +14,15 @@
   Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 */
 
-// This is a port of the arduino core file Print.h. see above notice
+// This is a port of the built-in arduino core file Print.h. see above notice , G Lyons
  
 #ifndef Print_h
 #define Print_h
 
 #include <inttypes.h>
 #include <stdio.h> // for size_t
-#include <string.h>
+#include <cstring> // for strlen
+#include <string>
 
 #define PGM_P const char*
 
@@ -68,7 +69,8 @@ class Print
     size_t print(long, int = DEC);
     size_t print(unsigned long, int = DEC);
     size_t print(double, int = 2);
-
+    size_t print(const std::string &);
+     
     size_t println(const char[]);
     size_t println(char);
     size_t println(int, int = DEC);
@@ -77,7 +79,8 @@ class Print
     size_t println(unsigned long, int = DEC);
     size_t println(double, int = 2);
     size_t println(void);
-
+    size_t println(const std::string &s);
+    
 };
 
 #endif