+++ extras/doc/CHANGELOG.md
@@ -6,3 +6,7 @@
 	* added print class for numerical Data
 * Version 1.2 April 2022
 	* Added support for 20x04 after part arrived and tested
+* Version 1.3 Jan 2023
+	* Added Example file for 20x04 LCD
+	* Added Example file for clock demo see issue one on github
+	* Added ability to change I2C bus speed in the constructor.